Updated project so it actually runs
This is a very large commit, but the goal of this was to make this
project be downloadable and work almost right out of the box. There
were prior sporadic updates that were not fully tested, and I wanted to
make sure anyone trying this out could get it up and pushed to a phone
with ease.

All of the AndroidBits code that was referenced before is now in its
own github repo (see README), so I had to do some refactoring due to
name changes.

Contributors, License, and Readme were all updated to reflect the new
status of the project.

Some icons and xml files were moved into different locations.

This commit has only tested the basic interface of the App, and it has
not been tested whether this version correctly forwards SMS's and
communicates properly with a server.

Todo: Fix anything preventing full functionality